% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_dataframe.R
\name{as_factor_dataframe}
\alias{as_factor_dataframe}
\title{as_factor_dataframe}
\usage{
as_factor_dataframe(df = NULL, columns = NULL)
}
\arguments{
\item{df}{A dataframe to perform the operation on}

\item{columns}{The columns you wish to convert to factors}
}
\description{
Converts columns in a dataframe to factors
}
