% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results_df.R
\docType{data}
\name{model_results_df}
\alias{model_results_df}
\title{model_results_df}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 2 columns.
}
\usage{
model_results_df
}
\description{
A toy dataset of model results
}
\seealso{
Other tutorial: 
\code{\link{as_factor_dataframe}()}
}
\concept{tutorial}
\keyword{datasets}
